+++ Sparky-core/src/maths/maths_func.h
@@ -7,12 +7,12 @@ namespace sparky { namespace maths {
 
 	inline float toRadians(float degrees)
 	{
-		return degrees * (M_PI / 180.0f);
+		return (float)(degrees * (M_PI / 180.0));
 	}
 
 	inline float toDegrees(float radians)
 	{
-		return radians * (180.0f / M_PI);
+		return (float)(radians * (180.0f / M_PI));
 	}
 
 	inline int sign(float value)
